Better error handling for API
- return 400 on undefined input
- return 404 on invalid id
- also fixed spacing

+++ functions/index.js
@@ -5,7 +5,7 @@ admin.initializeApp();
 const DEFAULT_LIMIT = 5
 const DEFAULT_EXPIRY = 60*60*1000;
 const MAX_EXPIRY = 30*24*60*60*1000;
-const decr = admin.firestore.FieldValue.increment(-1);
+const DECREMENT = admin.firestore.FieldValue.increment(-1);
 
 /**
 * Stores a new secret along with expiry and accesses left
@@ -14,39 +14,59 @@ const decr = admin.firestore.FieldValue.increment(-1);
 *	- secret
 */
 exports.create = functions.https.onRequest(async (req, res) => {
-  const secret = req.query.secret;
-  const limit = Number(req.query.limit) || DEFAULT_LIMIT;
-  const expiry = Number(req.query.expiry) || DEFAULT_EXPIRY; // in ms
+	const secret = req.query.secret;
+	const limit = Number(req.query.limit) || DEFAULT_LIMIT;
+	const expiry = Number(req.query.expiry) || DEFAULT_EXPIRY; // in ms
+	
+	if (typeof secret !== "string" || 
+			limit <= 0 || 
+			expiry <= 0 || 
+			expiry > MAX_EXPIRY) {
+		return res.sendStatus(400);
+	}
 
-  if (limit <= 0 || expiry <= 0 || expiry > MAX_EXPIRY) return res.sendStatus(400);
-
-  const expiryTime = Date.now() + expiry;
-
-  const writeResult = await admin.firestore().collection('messages').add({secret: secret, accessesLeft: limit, expiryTime: expiryTime});
-  res.json({id: `${writeResult.id}`});
+	const expiryTime = Date.now() + expiry;
+	const writeResult = await admin.firestore()
+  									.collection('messages')
+  									.add({	
+  											secret: secret, 
+  											accessesLeft: limit, 
+  											expiryTime: expiryTime
+  										});
+  
+  	res.json({id: `${writeResult.id}`});
 });
 
 /**
-* Fetches a secret by id. Returns "invalid id" if secret doesn't exist or expired or access limit reached.
+* Fetches a secret by id. Returns 404 if secret doesn't exist or expired or access limit reached.
 * Side effect: decrements accessesLeft.
 *
 * params:
 *	- id
 */
 exports.fetch = functions.https.onRequest(async (req, res) => {
-  const id = req.query.id;
-  const secret = await admin.firestore().collection('messages').doc(id).get();
+	const id = req.query.id;
 
-  if (!secret || !secret.exists) return res.json({secret: "invalid id"});
+	if (typeof id !== "string") {
+		return res.sendStatus(400);
+	}
 
-  const data = secret.data();
-  const isExpired = Date.now() > data.expiryTime;
+	const secret = await admin.firestore().collection('messages').doc(id).get();
 
-  if (data.accessesLeft <= 0 || isExpired) return res.json({secret: "invalid id"});
+	if (!secret || !secret.exists) {
+		return res.sendStatus(404);
+	}
 
-  decrAccessesLeft(id);
+	const data = secret.data();
+	const isExpired = Date.now() > data.expiryTime;
 
-  res.json({secret: secret.data().secret});
+	if (data.accessesLeft <= 0 || isExpired) {
+		return res.sendStatus(404);
+	}
+
+	decrementAccessesLeft(id);
+
+	res.json({secret: secret.data().secret});
 });
 
 /**
@@ -56,12 +76,16 @@ exports.fetch = functions.https.onRequest(async (req, res) => {
 *	- id
 */
 exports.delete = functions.https.onRequest(async (req, res) => {
-  const id = req.query.id;
-  const result = admin.firestore().collection('messages').doc(id).delete();
-  res.json({result: "ok"});
+	const id = req.query.id;
+
+	if (typeof id !== "string") {
+		return res.sendStatus(400);
+	}
+	
+	const result = admin.firestore().collection('messages').doc(id).delete();
+	res.json({result: "ok"});
 });
 
-function decrAccessesLeft(id) {
-	admin.firestore().collection('messages').doc(id).update({accessesLeft: decr})
+function decrementAccessesLeft(id) {
+	admin.firestore().collection('messages').doc(id).update({accessesLeft: DECREMENT})
 }
-
